% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{bifa_extdata}
\alias{bifa_extdata}
\title{Example data shipped in extdata}
\description{
This package ships example data files under \code{inst/extdata/}.
Use the loader functions below to access them.
}
\details{
Because these are stored in \code{extdata/}, they are not available via \code{data()}.
}
