% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsfa.R
\name{fit_bmsfa}
\alias{fit_bmsfa}
\title{Fit BMSFA via MSFA::sp_msfa}
\usage{
fit_bmsfa(
  Y_list,
  k,
  j_s,
  outputlevel = 1,
  scaling = FALSE,
  centering = TRUE,
  control = list(nrun = 3000, burn = 2000),
  ...
)
}
\arguments{
\item{Y_list}{List of study matrices.}

\item{k}{Number of common factors.}

\item{j_s}{Study-specific factors (length S).}

\item{outputlevel, scaling, centering, control}{Passed to \code{MSFA::sp_msfa()}.}

\item{...}{Passed to \code{MSFA::sp_msfa()}.}
}
\value{
A \code{bifa_fit} object.
}
\description{
Fit BMSFA via MSFA::sp_msfa
}
