% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsfa.R
\name{fit_bmsfa_2step}
\alias{fit_bmsfa_2step}
\title{Fit BMSFA all-at-once
If \code{post_fit0} is NULL, fit an initial BMSFA with given \code{k} and \code{j_s}, then automatically select
the number of factors and refit.
Otherwise, use the \code{post_fit0} then select the number of factors and refit.}
\usage{
fit_bmsfa_2step(
  Y_list,
  post_fit0 = NULL,
  k = NULL,
  j_s = NULL,
  cutoff = 0.05,
  ...
)
}
\arguments{
\item{Y_list}{Data list.}

\item{post_fit0}{Post-processed initial fit (or NULL).}

\item{k}{Initial number of common factors (if \code{post_fit0} is NULL).}

\item{j_s}{Initial study-specific factors (if \code{post_fit0} is NULL).}

\item{cutoff}{Eigen proportion cutoff.}

\item{...}{Passed to \code{fit_bmsfa()}.}
}
\description{
Fit BMSFA all-at-once
If \code{post_fit0} is NULL, fit an initial BMSFA with given \code{k} and \code{j_s}, then automatically select
the number of factors and refit.
Otherwise, use the \code{post_fit0} then select the number of factors and refit.
}
