% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_fa.R
\name{fit_ind_fa}
\alias{fit_ind_fa}
\title{Fit independent factor analysis (Ind FA) via MSFA::sp_fa (per study)}
\usage{
fit_ind_fa(
  Y_list,
  j_s,
  scaling = FALSE,
  centering = TRUE,
  control = list(nrun = 3000, burn = 2000),
  ...
)
}
\arguments{
\item{Y_list}{List of study matrices.}

\item{j_s}{Either a single integer or a length-S integer vector of factors.}

\item{scaling, centering, control}{Passed to \code{MSFA::sp_fa()}.}

\item{...}{Passed to \code{MSFA::sp_fa()}.}
}
\value{
A \code{bifa_fit} object whose \verb{$fit} is a list of fits (one per study).
}
\description{
Fit independent factor analysis (Ind FA) via MSFA::sp_fa (per study)
}
