% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_fa.R
\name{fit_ind_fa_2step}
\alias{fit_ind_fa_2step}
\title{Fit Ind FA all at once
If \code{post_fit0} is NULL, fit an initial IndFA with given \code{j_s}, then automatically select
the number of factors and refit.}
\usage{
fit_ind_fa_2step(Y_list, post_fit0 = NULL, j_s = NULL, cutoff = 0.05, ...)
}
\arguments{
\item{Y_list}{Data list.}

\item{post_fit0}{Initial post-processing result (or NULL to compute internally).}

\item{j_s}{Initial number of factors (or NULL to use all).}

\item{cutoff}{Eigen proportion cutoff.}

\item{...}{Passed to \code{fit_ind_fa()}.}
}
\description{
Fit Ind FA all at once
If \code{post_fit0} is NULL, fit an initial IndFA with given \code{j_s}, then automatically select
the number of factors and refit.
}
