% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{fit_integrative_fa}
\alias{fit_integrative_fa}
\title{Fit an integrative factor model with a chosen method within a general interface}
\usage{
fit_integrative_fa(method, ...)
}
\arguments{
\item{method}{One of: "bmsfa", "momss", "sufa", "pfa", "tetris", "stack_fa", "ind_fa".}

\item{...}{Passed to the corresponding \verb{fit_*()} function.}
}
\description{
Fit an integrative factor model with a chosen method within a general interface
}
