% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fa.R
\name{fit_stack_fa_2step}
\alias{fit_stack_fa_2step}
\title{Fit Stack FA all-at-once
If \code{post_fit0} is NULL, fit an initial StackFA with given \code{k}, then automatically select
the number of factors and refit.}
\usage{
fit_stack_fa_2step(Y_list, post_fit0 = NULL, k = NULL, cutoff = 0.05, ...)
}
\arguments{
\item{Y_list}{Data list.}

\item{post_fit0}{Initial post-processed result (or NULL to compute internally).}

\item{k}{Initial number of factors (used only if \code{post_fit0} is NULL).}

\item{cutoff}{Eigen proportion cutoff.}

\item{...}{Passed to \code{fit_stack_fa()}.}
}
\description{
Fit Stack FA all-at-once
If \code{post_fit0} is NULL, fit an initial StackFA with given \code{k}, then automatically select
the number of factors and refit.
}
