% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufa.R
\name{fit_sufa}
\alias{fit_sufa}
\title{Fit SUFA}
\usage{
fit_sufa(Y_list, k = NULL, nrun = 3000, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{Y_list}{List of study matrices.}

\item{k}{Positive integer for maximum number of factors shared (\code{qmax} in SUFA).}

\item{nrun}{Number of MCMC iterations.}

\item{center}{Whether to center each study matrix (default TRUE).}

\item{scale}{Whether to scale each study matrix (default FALSE).}

\item{...}{Passed to \code{SUFA::fit_SUFA()}.}
}
\value{
A \code{bifa_fit} object.
}
\description{
Thin wrapper around \code{SUFA::fit_SUFA()}
}
