% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{install_backend}
\alias{install_backend}
\title{Install optional backends used in the tutorial}
\usage{
install_backend(backend, ...)
}
\arguments{
\item{backend}{One of: "bmsfa", "momss", "sufa", "curated_ovarian".}

\item{...}{Passed to \code{remotes::install_github()} or \code{BiocManager::install()} when applicable.}
}
\description{
This is a convenience wrapper for non-CRAN dependencies.
}
\details{
\itemize{
\item For GitHub packages, we call \code{remotes::install_github()}.
\item For Bioconductor packages, we call \code{BiocManager::install()}.
\item For script-based methods (PFA / Tetris), scripts are bundled under \verb{inst/extdata/}.
}
}
