% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_toy}
\alias{load_toy}
\title{Load the toy example object}
\usage{
load_toy()
}
\value{
The R object stored in \code{toy.rds}. Often a list containing toy datasets.
}
\description{
Loads the file \code{inst/extdata/toy.rds}.
}
\examples{
toy <- load_toy()
str(toy)
}
