% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_fa.R
\name{postprocess_ind_fa}
\alias{postprocess_ind_fa}
\title{Post-process Ind FA output (nutrition tutorial)}
\usage{
postprocess_ind_fa(fit)
}
\arguments{
\item{fit}{A \code{bifa_fit} from \code{fit_ind_fa()} or a raw list of \code{MSFA::sp_fa()} fits.}
}
\value{
A list with \code{LambdaList}, \code{SigmaLambdaList}, \code{Psi} (list), \code{SigmaMarginal} (list).
}
\description{
Post-process Ind FA output (nutrition tutorial)
}
