% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{postprocess_integrative_fa}
\alias{postprocess_integrative_fa}
\title{Post-process an integrative factor model fit}
\usage{
postprocess_integrative_fa(method, fit, ...)
}
\arguments{
\item{method}{Backend name.}

\item{fit}{A fit object from the corresponding \verb{fit_*()} function.}

\item{...}{Passed to the corresponding \verb{postprocess_*()} function.}
}
\description{
Post-process an integrative factor model fit
}
