% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufa.R
\name{postprocess_sufa}
\alias{postprocess_sufa}
\title{Post-process SUFA output}
\usage{
postprocess_sufa(fit)
}
\arguments{
\item{fit}{A \code{bifa_fit} from \code{fit_sufa()} or a raw SUFA fit.}
}
\value{
A list with \code{Phi}, \code{SigmaPhi}, \code{SigmaMarginal}, \code{PsiList}.
}
\description{
Post-process SUFA output
}
