% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsfa.R
\name{select_js_bmsfa}
\alias{select_js_bmsfa}
\title{Select J_s for BMSFA from study-specific covariances}
\usage{
select_js_bmsfa(post, cutoff = 0.05)
}
\arguments{
\item{post}{Output of \code{postprocess_bmsfa()}.}

\item{cutoff}{Eigen proportion cutoff.}
}
\description{
Select J_s for BMSFA from study-specific covariances
}
