% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_fa.R
\name{select_js_ind_fa}
\alias{select_js_ind_fa}
\title{Select J_s for Ind FA (per-study) using eigen proportion rule}
\usage{
select_js_ind_fa(post, cutoff = 0.05)
}
\arguments{
\item{post}{Output of \code{postprocess_ind_fa()}.}

\item{cutoff}{Proportion threshold.}
}
\description{
Select J_s for Ind FA (per-study) using eigen proportion rule
}
