% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmsfa.R
\name{select_k_bmsfa}
\alias{select_k_bmsfa}
\title{Select K for BMSFA from common covariance}
\usage{
select_k_bmsfa(post, cutoff = 0.05)
}
\arguments{
\item{post}{Output of \code{postprocess_bmsfa()}.}

\item{cutoff}{Eigen proportion cutoff.}
}
\description{
Select K for BMSFA from common covariance
}
