% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fa.R
\name{select_k_stack_fa}
\alias{select_k_stack_fa}
\title{Select K for Stack FA using eigen proportion rule}
\usage{
select_k_stack_fa(post, cutoff = 0.05)
}
\arguments{
\item{post}{Output of \code{postprocess_stack_fa()}.}

\item{cutoff}{Proportion threshold (default 0.05).}
}
\description{
Select K for Stack FA using eigen proportion rule
}
